feat: create order route with idempotency

+++ internal/models/models.go
@@ -1,8 +1,9 @@
 package models
 
-type Order struct{
-  Id int8 `json:"id"`
+type Order struct {
+  Id int `json:"id"`
   IdempotencyId string `json:"idempotency_id"`
   Name string `json:"name"`
   Price float32 `json:"price"`
 }
+
